requirements.txt
python-telegram-bot
requests
beautifulsoup4
schedule
psutil
transmission-rpc